requirements.txt
scikit-learn==1.4.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
textblob==0.17.1
Pillow==10.2.0
numpy==1.26.3
pandas==2.2.0
python-dotenv==1.0.1
Flask==3.0.1
gunicorn==21.2.0
Flask-CORS==4.0.0 